Refactor 4 components ( Hero, programmesSeaction, getInvolved and ourLegacy) and create donate component
