--- src/pages/signUp.js
@@ -1,21 +1,0 @@
-export default function OurInitiatives() {
-  return (
-    <div className="container mx-auto py-16">
-      <h1 className="text-5xl font-bold text-primary text-center">Join Us</h1>
-      <p className="text-xl text-gray-700 mt-8 leading-relaxed text-center">
-        Become part of a community making a lasting difference. Whether it’s
-        volunteering your time or sharing your skills, you can help us reach
-        more people in need.
-      </p>
-
-      <div className="mt-12 text-center">
-        <a
-          href="#"
-          className="bg-primary text-white px-8 py-4 rounded-lg font-semibold hover:bg-blue-700 transition"
-        >
-          Sign Up to Volunteer
-        </a>
-      </div>
-    </div>
-  );
-}
